并行下载，细节优化

+++ Pages/DownloadPage.xaml.cs
@@ -4,7 +4,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Net;
 using System.Threading;
-using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Navigation;
@@ -17,7 +16,7 @@ namespace GBCLV2.Pages
         public bool Succeeded { get; internal set; }
         public readonly AutoResetEvent DownloadComplete = new AutoResetEvent(false);
         private readonly CancellationTokenSource cts = new CancellationTokenSource();
-        private readonly DispatcherTimer timer = new DispatcherTimer() { Interval = new TimeSpan(2500000) };
+        private readonly DispatcherTimer timer = new DispatcherTimer() { Interval = new TimeSpan(5000000) };
 
         private List<DownloadInfo> Downloads;
         private string title;
@@ -32,14 +31,13 @@ namespace GBCLV2.Pages
 
         public DownloadPage(List<DownloadInfo> FilesToDownload, string Title)
         {
-            ServicePointManager.DefaultConnectionLimit = 256;
+            ServicePointManager.DefaultConnectionLimit = 128;
 
             title = Title;
             Downloads = FilesToDownload;
             total = FilesToDownload.Count;
 
             InitializeComponent();
-            cancle_btn.Click += (s, e) => CancleDownload();
 
             timer.Tick += (s, e) =>
             {
@@ -56,7 +54,6 @@ namespace GBCLV2.Pages
                 if(complete + failed == total)
                 {
                     IsDownloading = false;
-                    cancle_btn.IsEnabled = false;
                     timer.Stop();
                     cts.Dispose();
                     titleBox.Text = "下载结束，某些文件下载失败";
@@ -64,23 +61,23 @@ namespace GBCLV2.Pages
                     return;
                 }
 
-                if(Bytes > 262144L)
+                if(Bytes > 524288L)
                 {
-                    speedBox.Text = $"{Bytes / 262144.0:F1} MB/s";
+                    speedBox.Text = $"{Bytes / 524288.0:F1} MB/s";
                 }
-                else if(Bytes > 256L)
+                else if(Bytes > 512L)
                 {
-                    speedBox.Text = $"{Bytes / 256.0:F1} KB/s";
+                    speedBox.Text = $"{Bytes / 512.0:F1} KB/s";
                 }
                 else
                 {
-                    speedBox.Text = $"{Bytes * 4} B/s";
+                    speedBox.Text = $"{Bytes << 1} B/s";
                 }
                 Bytes = 0;
             };
         }
 
-        private async void StartDownload(object sender, RoutedEventArgs e)
+        private void StartDownload(object sender, RoutedEventArgs e)
         {
             IsDownloading = true;
             Succeeded = false;
@@ -90,24 +87,14 @@ namespace GBCLV2.Pages
             timer.Start();
             statusBox.Text = $"0/{total}个文件下载成功";
 
-            foreach (var download in Downloads)
+            foreach(var download in Downloads)
             {
-                if(cts.IsCancellationRequested) break;
-                await DownloadFileAsync(download).ConfigureAwait(false);
+                ThreadPool.QueueUserWorkItem(t => DownloadFile(download));
             }
+
         }
 
-        private void CancleDownload()
-        {
-            cts.Cancel();
-            timer.Stop();
-            titleBox.Text = "下载已取消";
-            speedBox.Text = null;
-            IsDownloading = false;
-            cancle_btn.IsEnabled = false;
-        }
-
-        private async Task DownloadFileAsync(DownloadInfo download)
+        private void DownloadFile(DownloadInfo download)
         {
             if (!Directory.Exists(Path.GetDirectoryName(download.Path)))
             {
@@ -118,13 +105,13 @@ namespace GBCLV2.Pages
             {
                 HttpWebRequest request = (HttpWebRequest)WebRequest.Create(download.Url);
                 request.Proxy = null;
-                request.Timeout = 6000;
-                request.ReadWriteTimeout = 18000;
+                request.Timeout = 10000;
+                request.ReadWriteTimeout = 25000;
                 request.Method = "GET";
                 request.UserAgent = "Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.0; .NET CLR 1.0.3705)";
 
                 cts.Token.Register(() =>request.Abort());
-                HttpWebResponse response = await request.GetResponseAsync() as HttpWebResponse;
+                HttpWebResponse response = request.GetResponse() as HttpWebResponse;
 
                 if (cts.IsCancellationRequested)
                 {
@@ -140,18 +127,12 @@ namespace GBCLV2.Pages
                     while (!cts.IsCancellationRequested && size > 0)
                     {
                         fileStream.Write(buffer, 0, size);
-                        size = await responseStream.ReadAsync(buffer, 0, 1024);
+                        size = responseStream.Read(buffer, 0, 1024);
                         Bytes += size;
                     }
                     responseStream.Close();
                     fileStream.Close();
                     response.Dispose();
-                }
-
-                if (cts.IsCancellationRequested)
-                {
-                    File.Delete(download.Path);
-                    return;
                 }
             }
             catch(WebException) when(cts.IsCancellationRequested)
@@ -162,9 +143,13 @@ namespace GBCLV2.Pages
             catch
             {
                 Interlocked.Increment(ref failed);
-                //Debug.WriteLine(download.Url);
-                //Debug.WriteLine(e.Message);
                 Dispatcher.Invoke(() => failsBox.Text = $"{failed}个文件下载失败" );
+                File.Delete(download.Path);
+                return;
+            }
+
+            if (cts.IsCancellationRequested)
+            {
                 File.Delete(download.Path);
                 return;
             }
@@ -177,14 +162,23 @@ namespace GBCLV2.Pages
             });
         }
 
+        private void CancleDownload()
+        {
+            titleBox.Text = "取消中";
+            speedBox.Text = null;
+            cts.Cancel();
+            timer.Stop();
+            IsDownloading = false;
+        }
+
         private void Go_Back(object sender, RoutedEventArgs e)
         {
             if(IsDownloading)
             {
                 if (MessageBox.Show("正在下载中，你确定要中止吗", "(●—●)", MessageBoxButton.OKCancel, MessageBoxImage.Asterisk) == MessageBoxResult.OK)
                 {
+
                     CancleDownload();
-                    cts.Dispose();
                     NavigationService.GoBack();
                     DownloadComplete.Set();
                 }
